refactor: Modularização
